Scraped annotated data for each of the 12 sectors using a python package named Newspaper. If you want to scrape additionally yourself, I believe you'll have to install the package. The instructions to do so can be found here: http://newspaper.readthedocs.org/en/latest/user_guide/install.html#install
These documents are in text files with the following formatting:
Article Title
Author(s)
URL
Article Content

It should also be noted that these articles were scraped from the 20 most recent pages for each of the sectors. Consequently, this annotated dataset will not extend very far back in time. While this will most likely not be an issue for the purposes of document classification, it is something to remember.